fix: pageinfo bug fix; refactor: swreg func
