+++ src/components/Header/Header.scss
@@ -51,6 +51,7 @@
   user-select: none;
   font-size: 15px;
   font-weight: $medium;
+  font-family: $appfont;
 }
 
 @media (max-width: $minimum-screen) {
