+++ src/styles.scss
@@ -23,6 +23,8 @@ $mobile-screen3: 650px;
 $mobile-screen2: 600px;
 $mobile-screen1: 500px;
 $minimum-screen: 320px;
+$appfont: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu',
+'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
 
 @keyframes swing {
   15% {
